feature: store stats last week to calculate delta data
